enable the rtsp-simple-server api
This change will enable the rtsp-simple-server api which will hopefully allow users to create a health check in the `docker-compose.yml`
I am hoping this endpoint:
 - https://aler9.github.io/rtsp-simple-server/#operation/pathsList
will enable docker to check if the streams are available and if not restart the container. 
If not I am hoping another one of the endpoints will work for checking the health of the streams.